:sparkles: Add repository create and update api (#123)

+++ pkg/handlers/repositories/repositories_delete.go
@@ -12,10 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package repository
+package repositories
 
 import (
 	"errors"
+	"fmt"
 	"net/http"
 
 	"github.com/labstack/echo/v4"
@@ -29,10 +30,16 @@ import (
 
 // DeleteRepository handles the delete repository request
 // @Summary Delete repository
+// @Tags Repository
+// @security BasicAuth
 // @Accept json
 // @Produce json
-// @Router /repository/{id} [delete]
+// @Router /namespaces/{namespace}/repositories/{id} [delete]
+// @Param namespace path string true "Namespace"
+// @Param id path string true "Repository ID"
 // @Success 204
+// @Failure 404 {object} xerrors.ErrCode
+// @Failure 500 {object} xerrors.ErrCode
 func (h *handlers) DeleteRepository(c echo.Context) error {
 	ctx := log.Logger.WithContext(c.Request().Context())
 
@@ -43,7 +50,32 @@ func (h *handlers) DeleteRepository(c echo.Context) error {
 		return xerrors.NewHTTPError(c, xerrors.HTTPErrCodeBadRequest, err.Error())
 	}
 
+	namespaceService := h.namespaceServiceFactory.New()
+	namespaceObj, err := namespaceService.GetByName(ctx, req.Namespace)
+	if err != nil {
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			log.Error().Err(err).Str("namespace", req.Namespace).Msg("Namespace not found")
+			return xerrors.NewHTTPError(c, xerrors.HTTPErrCodeNotFound, fmt.Sprintf("Namespace(%s) not found: %v", req.Namespace, err))
+		}
+		log.Error().Err(err).Str("namespace", req.Namespace).Msg("Namespace find failed")
+		return xerrors.NewHTTPError(c, xerrors.HTTPErrCodeInternalError, fmt.Sprintf("Namespace(%s) find failed: %v", req.Namespace, err))
+	}
+
 	repositoryService := h.repositoryServiceFactory.New()
+	repositoryObj, err := repositoryService.Get(ctx, req.ID)
+	if err != nil {
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			log.Error().Err(err).Int64("id", req.ID).Msg("Repository not found")
+			return xerrors.NewHTTPError(c, xerrors.HTTPErrCodeNotFound, err.Error())
+		}
+		log.Error().Err(err).Msg("Repository find failed")
+		return xerrors.NewHTTPError(c, xerrors.HTTPErrCodeInternalError, err.Error())
+	}
+	if repositoryObj.NamespaceID != namespaceObj.ID {
+		log.Error().Interface("repositoryObj", repositoryObj).Interface("namespaceObj", namespaceObj).Msg("Repository's namespace ref id not equal namespace id")
+		return xerrors.NewHTTPError(c, xerrors.HTTPErrCodeNotFound)
+	}
+
 	err = repositoryService.DeleteByID(ctx, req.ID)
 	if err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
